use same prs display in agent runs
